pull name_and_object_from_url to utils
