+++ lmfdb/lfunctions/Lfunctionutilities.py
@@ -8,6 +8,11 @@ from lmfdb.genus2_curves.web_g2c import list_to_factored_poly_otherorder
 from lmfdb.transitive_group import group_display_knowl
 from lmfdb.base import getDBConnection
 from lmfdb.utils import truncate_number
+from lmfdb.elliptic_curves.web_ec import is_ec_isogeny_class_in_db
+from lmfdb.ecnf.WebEllipticCurve import is_ecnf_isogeny_class_in_db
+from lmfdb.hilbert_modular_forms.web_HMF import is_hmf_in_db
+from lmfdb.bianchi_modular_forms.web_BMF import is_bmf_in_db
+from lmfdb.modular_forms.elliptic_modular_forms.backend.emf_utils import is_newform_in_db
 
 ###############################################################
 # Functions for displaying numbers in correct format etc.
@@ -64,7 +69,7 @@ def string2number(s):
     except:
         return s
 
-    
+
 def pair2complex(pair):
     ''' Turns the pair into a complex number.
     '''
@@ -121,7 +126,7 @@ def seriescoeff(coeff, index, seriescoefftype, seriestype, truncationexp, precis
     except TypeError:     # mostly a hack for Dirichlet L-functions
         if seriescoefftype == "serieshtml":
             if coeff == "I":
-                return " + " + "$i$" + "&middot;" + seriesvar(index, seriestype) 
+                return " + " + "$i$" + "&middot;" + seriesvar(index, seriestype)
             elif coeff == "-I":
                 return "&minus;" + " $i$" + "&middot;" + seriesvar(index, seriestype)
             else:
@@ -295,7 +300,7 @@ def lfuncDShtml(L, fmt):
 #    if fmt == "analytic" or fmt == "langlands":
     if fmt in ["analytic", "langlands", "arithmetic"]:
         ans += "<table class='dirichletseries'><tr>"
-        ans += "<td valign='top'>"  # + "$" 
+        ans += "<td valign='top'>"  # + "$"
         if fmt == "arithmetic":
             ans += "<span class='term'>"
             ans += L.htmlname_arithmetic
@@ -333,9 +338,9 @@ def lfuncDShtml(L, fmt):
                     "serieshtml", "dirichlethtml", -6, 5)
             if tmp != "":
                 nonzeroterms += 1
-            ans = ans + " <span class='term'>" + tmp + "</span> "  
+            ans = ans + " <span class='term'>" + tmp + "</span> "
                 # need a space between spans to allow line breaks. css stops a break within a span
-     
+
             if nonzeroterms > maxcoeffs:
                 break
             if(nonzeroterms % numperline == 0):
@@ -466,18 +471,18 @@ def lfuncEPhtml(L,fmt):
     for lf in L.bad_lfactors:
         try:
             thispolygal = list_to_factored_poly_otherorder(lf[1], galois=True)
-            eptable += ("<tr><td>" + goodorbad + "</td><td>" + str(lf[0]) + "</td><td>" + 
+            eptable += ("<tr><td>" + goodorbad + "</td><td>" + str(lf[0]) + "</td><td>" +
                         "$" + thispolygal[0] + "$" +
                         "</td>")
             if L.degree > 2:
-                eptable += "<td class='galois'>" 
+                eptable += "<td class='galois'>"
                 this_gal_group = thispolygal[1]
                 if this_gal_group[0]==[0,0]:
                     pass   # do nothing, because the local faco is 1
                 elif this_gal_group[0]==[1,1]:
-                    eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C,'$C_1$') 
+                    eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C,'$C_1$')
                 else:
-                    eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C) 
+                    eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C)
                 for j in range(1,len(thispolygal[1])):
                     eptable += "$\\times$"
                     eptable += group_display_knowl(this_gal_group[j][0],this_gal_group[j][1],C)
@@ -500,7 +505,7 @@ def lfuncEPhtml(L,fmt):
         if L.degree > 2:
             eptable += "<td class='galois'>"
             this_gal_group = thispolygal[1]
-            eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C) 
+            eptable += group_display_knowl(this_gal_group[0][0],this_gal_group[0][1],C)
             for j in range(1,len(thispolygal[1])):
                 eptable += "$\\times$"
                 eptable += group_display_knowl(this_gal_group[j][0],this_gal_group[j][1],C)
@@ -577,7 +582,7 @@ def lfuncEpSymPower(L):
         ans += poly_string
     ans += '\\prod_{p \\nmid %d }\\prod_{j=0}^{%d} ' % (L.E.conductor(),L.m)
     ans += '\\left(1- \\frac{\\alpha_p^j\\beta_p^{%d-j}}' % L.m
-    ans += '{p^{s}} \\right)^{-1}'    
+    ans += '{p^{s}} \\right)^{-1}'
     return ans
 
 #---------
@@ -889,7 +894,7 @@ def getConductorIsogenyFromLabel(label):
             while iso[0].isdigit():
                 cond += iso[0]
                 iso = iso[1:]
-                
+
         # Strip off the curve number
         while iso[-1].isdigit():
             iso = iso[:-1]
@@ -897,6 +902,44 @@ def getConductorIsogenyFromLabel(label):
 
     except:
         return None, None
-    
 
+# TODO This needs to be able to handle any sort of L-function.
+# There should probably be a more relevant field
+# in the database, instead of trying to extract this from a URL
+def name_and_object_from_url(url):
+    url_split = url.split("/");
+    name = None;
+    obj_exists = False;
 
+    if url_split[0] == "EllipticCurve":
+        if url_split[1] == 'Q':
+            # EllipticCurve/Q/341641/a
+            label_isogeny_class = ".".join(url_split[-2:]);
+            # count doesn't honor limit!
+            obj_exists = is_ec_isogeny_class_in_db(label_isogeny_class);
+        else:
+            # EllipticCurve/2.2.140.1/14.1/a
+            label_isogeny_class =  "-".join(url_split[-3:]);
+            obj_exists = is_ecnf_isogeny_class_in_db(label_isogeny_class);
+        name = 'Isogeny class ' + label_isogeny_class;
+
+    elif url_split[0] == "ModularForm":
+        if url_split[1] == 'GL2':
+            if url_split[2] == 'Q' and url_split[3]  == 'holomorphic':
+                # ModularForm/GL2/Q/holomorphic/14/2/1/a
+                full_label = ".".join(url_split[-4:])
+                name =  'Modular form ' + full_label;
+                obj_exists = is_newform_in_db(full_label);
+
+            elif  url_split[2] == 'TotallyReal':
+                # ModularForm/GL2/TotallyReal/2.2.140.1/holomorphic/2.2.140.1-14.1-a
+                label = url_split[-1];
+                name =  'Hilbert modular form ' + label;
+                obj_exists = is_hmf_in_db(label);
+
+            elif url_split[2] ==  'ImaginaryQuadratic':
+                # ModularForm/GL2/ImaginaryQuadratic/2.0.4.1/98.1/a
+                label = '-'.join(url_split[-3:])
+                name = 'Bianchi modular form ' + label;
+                obj_exists = is_bmf_in_db(label);
+    return name, obj_exists
